[chore]: Move sceneManager to Core folder
