+++ include/Core/SceneManager.hpp
